First Commit- A Complete project
This is a complete project with database. Basically this project is for a beginner . I could not follow the Proper structure for making project still you can understand how the application works. I had made the project when I had 0 year experience. To Know more how  To to run application please read the readme file.